Fix character gravity and add character model
With this commit the character will now keep from falling through the
map. I also added a model for the character and adjusted the camera view
to show him. The walls in the maze were made higher as well so keep from
him jumping over them.

There is an issue with being able to see through walls by standing too
close, but that will be addressed later.