Update character & map models
Updated the character camera to now be a first person camera
instead of a third person camera.

Edited the character models offset to make sure the player
cannot see through walls.

Lowered the characters movement speed, because once the camera was
in first person it was much harder to control. It is now set to what
would be considered a brisk walking pace.